+++ buffalogs/impossible_travel/views.py
@@ -131,13 +131,13 @@ def get_users(request):
 def get_all_logins(request, pk_user):
     context = []
     count = 0
-    connections.create_connection(hosts=[settings.CERTEGO_ELASTICSEARCH], timeout=90)
+    connections.create_connection(hosts=[settings.CERTEGO_BUFFALOGS_ELASTICSEARCH], timeout=90)
     end_date = timezone.now()
     start_date = end_date + timedelta(days=-365)
     user_obj = User.objects.filter(id=pk_user)
     username = user_obj[0].username
     s = (
-        Search(index=settings.CERTEGO_ELASTIC_INDEX)
+        Search(index=settings.CERTEGO_BUFFALOGS_ELASTIC_INDEX)
         .filter("range", **{"@timestamp": {"gte": start_date, "lt": end_date}})
         .query("match", **{"user.name": username})
         .exclude("match", **{"event.outcome": "failure"})