+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 1.2.x
 ### 1.2.0
 #### Features
-* Implemented Certego shared CI 1.3.5 
+* Implemented Certego shared CI 1.3.5
+#### Changes
+* Renamed constants
 
 ## 1.1.x
 ### 1.1.1